:bug: Fix eleven defects found reviewing the five longest modules
Code review of remesh, holes, kernels/remesh, smoothing and creation.
Every fix is verified against a detached baseline worktree or by a
mutation probe naming the assertion that fails without it.

Correctness:

- The collapse lock key was not injective. `lowbias32(k) & 0x7FFFFFFF`
  is exactly 2-to-1 and `wins_key_everywhere` tests equality against a
  neighbourhood minimum, so two candidates could both commit into
  overlapping 1-rings. The quadric path guarded this with a second lock
  pass; the isotropic path did not. Making the key 64-bit injective
  (hash high, index low) closes it for both and retires
  `claim_collapse_index` and one launch per round.
- `emit_unique_edges` clamped the row write but not `out_inverse`, the
  index `scatter_edge_incidence` dereferences -- so the guard against an
  out-of-bounds write left an out-of-bounds read.
- `collapse_candidates`' anti-oscillation walked only the removed
  endpoint's ring; a free placement moves the survivor too. Costs ~4% of
  collapses on a graded saddle, aspect ratios unchanged.
- `stitch_dp_diag` read `b_opp[j % n_b]` where the rim edge it
  introduces lives at slot `j - 1`.
- `fillable_loop_mask` returned True for a loop with a chord: vertex
  ownership was last-write-wins across loops sharing a vertex.
- `_bridge_triangles` emitted an empty patch at `a0 == b0` and a doubled
  one at `a1 == b1`. Both are reachable through `bridge_edges`.
- `refine` went unvalidated on the `triangulate_only` and no-op paths.
- `pin_boundary=True` on a mesh with no interior vertex ran the
  *unconstrained* flow: the patch collapsed to a point.
- The Dirichlet solve seeded CG from the right-hand side, teleporting
  unreferenced vertices to the origin.
- A signed-volume sign flip made the volume constraint raise TypeError.
- `uv_sphere` mutated the caller's `count` array, and resolution 2 on a
  wrapped axis silently produced a zero-face mesh.

Performance, bit-identical output (interleaved, min of 9, 3 reps):

- `filter_implicit_fairing` 27.6 -> 13.2 ms (2.06-2.12x) and
  `filter_laplacian(implicit)` 19.7 -> 6.9 ms (2.77-2.83x), by batching
  the three position components through `solve_spd_columns` instead of
  three separate single-column solves.

Reuse:

- `subdivide_region_to_size` calls `split_edges`; `intrinsic_delaunay`
  uses `_flip_interior_edges` through a `before_commit` hook (verified
  identical flip counts, faces and lengths).
- Shared funcs for the link condition and the cell-centre distance;
  merged the upload helpers and the cylinder/annulus axis block.
- Dropped `_EdgeIncidence.inverse`, an unreachable `capture=True`
  branch, a redundant readback, two discarded buffers and a duplicate
  scan.

Gates: 3153 passed / 1 skipped, both-device runner green, ruff clean,
basedpyright 0 errors, mkdocs --strict clean, parity 104 pairs unchanged.

+++ triwarp/kernels/scatter.py
@@ -198,18 +198,18 @@ def lock_two_rings(
     columns: wp.array[wp.int32],
     s: wp.int32,
     r: wp.int32,
-    key: wp.int32,
-    out_claim: wp.array[wp.int32],
+    key: wp.int64,
+    out_claim: wp.array[wp.int64],
 ) -> None:
     # Atomic-min ``key`` into every vertex of the two closed 1-rings of ``s`` and ``r``: the lock
     # half of a parallel independent set over edge candidates, so that whichever candidate wins
     # everywhere it touched has a neighbourhood disjoint from every other winner's.
     #
-    # ``kernels/remesh.py`` runs this three times per collapse pass -- once for the isotropic
-    # path's single claim and twice for the quadric path's key-then-index pair -- with a different
-    # ``key`` each time, which is the whole of the difference between the three. What the *key*
-    # must be is recorded at ``remesh.scramble_index``, and it is load-bearing: a spatially
-    # monotone key commits one collapse per pass.
+    # Both of ``kernels/remesh.py``'s collapse paths run this once per pass, with the key
+    # ``remesh.scramble_index`` builds. Two properties of that key are load-bearing and both are
+    # recorded there: it must not be spatially monotone (a monotone key commits one collapse per
+    # pass), and it must be *injective*, or two candidates can tie and both believe they won --
+    # which is why it is 64 bits wide here rather than the natural int32 of a vertex index.
     wp.atomic_min(out_claim, s, key)
     wp.atomic_min(out_claim, r, key)
     for i in range(offsets[s], offsets[s + 1]):
